+++ tests/Installer/Writer/PhpCsConfigWriterTest.php
@@ -39,7 +39,7 @@ class PhpCsConfigWriterTest extends TestCase
  * Additional rules or rules to override.
  * These rules will be added to default rules or will override them if the same key already exists.
  */
- 
+
 $additionalRules = [];
 $rulesProvider = new Facile\CodingStandards\Rules\CompositeRulesProvider([
     new Facile\CodingStandards\Rules\DefaultRulesProvider(),
@@ -83,7 +83,7 @@ TEXT;
  * Additional rules or rules to override.
  * These rules will be added to default rules or will override them if the same key already exists.
  */
- 
+
 $additionalRules = [];
 $rulesProvider = new Facile\CodingStandards\Rules\CompositeRulesProvider([
     new Facile\CodingStandards\Rules\DefaultRulesProvider(),
@@ -127,7 +127,7 @@ TEXT;
  * Additional rules or rules to override.
  * These rules will be added to default rules or will override them if the same key already exists.
  */
- 
+
 $additionalRules = [];
 $rulesProvider = new Facile\CodingStandards\Rules\CompositeRulesProvider([
     new Facile\CodingStandards\Rules\DefaultRulesProvider(),
@@ -170,7 +170,7 @@ TEXT;
  * Additional rules or rules to override.
  * These rules will be added to default rules or will override them if the same key already exists.
  */
- 
+
 $additionalRules = [];
 $rulesProvider = new Facile\CodingStandards\Rules\CompositeRulesProvider([
     new Facile\CodingStandards\Rules\DefaultRulesProvider(),